First Commit. Everything is in place except for the airplane input. Fix some bugs and test.

+++ src/java_proj_group_7/ParkingLot.java
@@ -3,7 +3,7 @@ package java_proj_group_7;
 public class ParkingLot {
 	
 	private boolean isEmpty = true; 
-	private int nextAvailible = 0; 
+	private long nextAvailible = 0; 
 	
 	public boolean isEmpty() {
 		return isEmpty;
@@ -13,12 +13,12 @@ public class ParkingLot {
 		isEmpty = true;
 	} 
 	
-	public void occupyParkingLot(int time) { 
+	public void occupyParkingLot(long time) { 
 		isEmpty = false; 
 		this.nextAvailible = time;
 	}
 
-	public int getNextAvailible() {
+	public long getNextAvailible() {
 		return nextAvailible;
 	}
